Complete test for 'compose' function factory: split into submodules, complete fixtures
