+++ tests/fixtures/functional.py
@@ -73,21 +73,6 @@ def maps() -> Sequence[Map]:
 
 
 @pytest.fixture(scope='function')
-def suitable_maps() -> Sequence[Map]:
-    return find(strategies.suitable_maps)
-
-
-@pytest.fixture(scope='function')
-def other_suitable_maps() -> Sequence[Map]:
-    return find(strategies.suitable_maps)
-
-
-@pytest.fixture(scope='function')
-def another_suitable_maps() -> Sequence[Map]:
-    return find(strategies.suitable_maps)
-
-
-@pytest.fixture(scope='function')
 def maps_arguments(maps: Sequence[Map]) -> Sequence[Domain]:
     maps_count = len(maps)
     return find(strategies.to_homogeneous_sequences(strategies.maps_arguments,
@@ -108,6 +93,21 @@ def last_map(next_map: Map[Range, Intermediate]) -> Map[Intermediate, Range]:
 @pytest.fixture(scope='function')
 def suitable_maps() -> Tuple[Map, ...]:
     return find(strategies.suitable_maps)
+
+
+@pytest.fixture(scope='function')
+def various_suitable_maps() -> Tuple[Map, ...]:
+    return find(strategies.various_suitable_maps)
+
+
+@pytest.fixture(scope='function')
+def other_various_suitable_maps() -> Sequence[Map]:
+    return find(strategies.various_suitable_maps)
+
+
+@pytest.fixture(scope='function')
+def another_various_suitable_maps() -> Sequence[Map]:
+    return find(strategies.various_suitable_maps)
 
 
 @pytest.fixture(scope='function')
