+++ tests/strategies/__init__.py
@@ -19,7 +19,8 @@ from .functional import (false_predicates,
                          to_transparent_functions_args,
                          to_transparent_functions_kwargs,
                          transparent_functions,
-                         true_predicates)
+                         true_predicates,
+                         various_suitable_maps)
 from .iterating import (groupers_keys,
                         non_negative_indices,
                         non_negative_slices)
